Module rustcode::data_structures::concurrency

A module for concurrency primitives

Reexports

use core::prelude::v1::*;

use core::prelude::v1::*;

use core::sync::atomic::{AtomicIsize, Ordering};

use core::cell::UnsafeCell;

use core::option::Option::Some;

use core::ops::{Deref, DerefMut};

use alloc::boxed::Box;

use super::super::ProcessQueue;

use super::super::super::process::{ready_queue, proc_yield};

use super::super::super::interrupts::{on, off};

Modules

semaphore
A semaphore implementation based on the rust Mutex type

Structs

Semaphore
Semaphore is a much more Rustic semaphore. It returns an RAII SemaphoreGuard, which automatically calls "up" when it goes out of scope. This semaphore takes ownership of the data it is guarding, so that Rust ownership and lifetime semantics can be used to guarantee safety of the resource.

SemaphoreGuard
RAII SemaphoreGuard

StaticSemaphore
StaticSemaphore is a semaphore implementation that can be used in statics. It has a const constructor.