Struct rustcode::data_structures::proc_queue::ProcessQueue

pub struct ProcessQueue {
    pub head: *mut Process,
    pub tail: *mut Process,
}

An abstraction for any kind of process queue

This structure should be used whenever queueing up processes. It uses raw pointers because it is often necessary to have multiple pointers to a process with atomic access to the process (as opposed to making a function call to get a pointer to the process).

It is uses the pointer already present in the Process struct to manage chains of processes.

This is not very Rustic at all :( but it cannot be helped because Rust's abstractions are based on the OS in many cases.

Because of this, the queue does not do any memory management at all!

Fields

head

tail

Methods

impl ProcessQueue

const fn new() -> ProcessQueue

Create a new, empty queue. This is a const function, so statics can have queues.

fn push_tail(&mut self, process: *mut Process)

Push the given process to the end of the queue. O(1)

fn pop_head(&mut self) -> *mut Process

Pop and return the head of the queue. If this method returns a null pointer, then the queue is empty. O(1)