Module rustcode::memory::heap

This file contains the memory allocator used by the rust_alloc module

***INTERRUPTS MUST BE OFF BEFORE RUNNING ANYTHING IN THIS FILE!!!!

The implementation in this file is a simple first-fit allocator.

Invariants: * BLOCK_ALIGN = 4*size_of::() * All blocks will be a multiple of BLOCK_ALIGN * All blocks will be BLOCK_ALIGN-aligned

Block structure: * size * forward pointer | free bits * backward pointer | 0x0 ... * size (last word)

When a block is free, the first and last words should match and be equal to the size of the block. The forward pointer points to the head of the next free block. The backward pointer points to the head of the previous free block. Since all blocks are at least 16B aligned, at least the last 4 bits of all block addresses are 0. The last 4 bits of the forward pointer should be all 1s if the block is free.

When a block is in use, the whole block is usable for the user. Thus, usable size and block size are equal.

Reexports

use core::prelude::v1::*;

use core::mem::{size_of};

Structs

Block
A struct representing a heap block.

Constants

DEBUG
A flag to turn on and off debugging output

Statics

BLOCK_ALIGN
A const representing the minimum alignment of any block. It is effectively a const, but it has to be initialized at startup because size_of is non-const.

END
The end address of the kernel heap. That is, the first address that is not in the heap.

FAIL_MALLOCS
The number of unsuccessful mallocs (for stats purposes)

FREES
The number of successful frees (for stats purposes)

START
The start address of the kernel heap

SUCC_MALLOCS
The number of successful mallocs (for stats purposes)

free_list
A pointer to the first free heap block

Functions

free
Deallocates the memory referenced by ptr.

get_block_stats
Helper method to compute stats

init
Initialize the kernel heap

malloc
Return a pointer to size bytes of memory aligned to align.

print_stats
Prints implementation-defined allocator statistics.

round_to_block_align
Round up to the nearest multiple of BLOCK_ALIGN

round_to_n
Round the size to n. n must be a power of 2

usable_size
Returns the usable size of an allocation created with the specified the size and align.