Struct rustcode::memory::regionmap::Region

struct Region {
    start: usize,
    end: usize,
    usable: bool,
    next: Option<Box<Region>>,
}

Represents a region of memory

Fields

start

end

usable

next

Methods

impl Region

fn new(start: usize) -> Region

Recursively create regions and return a chain that stretches from start to the end of the 32-bit addr space. Each iteration is O(n), but it doesn't matter much because n < 2*20

fn find(addr: usize) -> Option<(usize, usize)>

Find the length and type of the most restrictive mapping for addr based on the E820 BIOS call

fn find_next(addr: usize) -> usize

Find the beginning address of the next region of memory based on the E820 BIOS call

fn dump(&self)

Recursively print mappings. Used for debugging