Struct rustcode::memory::AddressSpace

pub struct AddressSpace {
    page_dir: usize,
    kmap_index: u8,
}

The address space of a single process

Fields

page_dir

kmap_index

Methods

impl AddressSpace

fn new() -> AddressSpace

Create a new address space. Each address space has the first 8MiB direct mapped. The next 4MiB is mapped to the page directory

fn map(&mut self, phys: usize, virt: usize)

Map virt to phys in the address space. NOTE: should only be called on the current address space because it assumes that the PD is at 0x802000

fn kmap(&mut self, paddr: usize) -> &mut Frame

Map the given paddr for temporary use by the kernel and return a mut reference to the frame. NOTE: should only be called on the current address space because it assumes that the PD is at 0x802000

fn unmap(&mut self, virt: usize)

Remove any mapping for this virtual address NOTE: should only be called on the current address space because it assumes that the PD is at 0x802000

fn activate(&mut self)

Activate the current address space and turn on VM if needed

fn clear(&mut self)

Remove all non-kernel mappings in this address space. NOTE: must run while this address space is active