Module rustcode::memory::virtmem

A module for virtual memory management The virtual address space of a process consists of all addresses for 0xD0_0000-0xFFFF_FFFF. When a page fault occurs (in kernel or user mode), if the faulting address is in this range, the fault is considered an error, and the process is killed. This is because the first 8MiB are shared, direct-mapped memory and should never produce a page fault. The next 4MiB are mapped to the processes page directory, so if there is a fault in this range, it is fatal, since it means the paging structures have been corrupt. The 13MiB is claimed by the kernel for use by the kmap function, which temporarily maps a page meant to reside in another process's address space. This is needed for creating a new process. Luckly, with Rust, segfaults should be extremely rare.

When a page fault occurs for a legal address, the fault handler checks the following cases:

Is the page paged out to disk? If so, swap it back in

Is the page marked present but read-only? If so, this page is COW, so clone it and mark the new page read/write

Else, allocate a new frame a map the page to it

Reexports

use core::prelude::v1::*;

use core::ops::{Index, IndexMut};

use core::intrinsics::transmute;

use super::physmem::Frame;

use super::super::machine::{invlpg, vmm_on, page_fault_handler};

use super::super::interrupts::{add_trap_handler, on, off};

use super::super::process::CURRENT_PROCESS;

Structs

AddressSpace
The address space of a single process

PagingEntry
A single entry in a page directory or table 31 12 9 0 [ 000 flags ]

VMTable
An abstraction of page directories and page tables

Statics

PDE0
Shared PDE #0: direct mapping the first 4MiB

PDE1
Shared PDE #1: direct mapping the second 4MiB

VMM_ON
Is VM on?

Functions

init
Initialize virtual memory management but do not turn on VM. This creates the two shared page tables that map the first 8MiB.

vmm_page_fault
The Rust-side code of the page fault handler.