Module rustcode::process

A module for process management

Reexports

use core::prelude::v1::*;

use alloc::boxed::Box;

use core::option::Option::{self, Some, None};

use core::fmt::{Debug, Formatter, Result};

use core::cmp::PartialEq;

use core::sync::atomic::{AtomicUsize, Ordering};

use core::ops::Drop;

use super::data_structures::ProcessQueue;

use super::interrupts::{on, off};

use super::machine::{self, context_switch};

use super::memory::AddressSpace;

use self::context::{KContext};

use self::idle::IDLE_PROCESS;

Modules

context
A module for saving a process's context

idle
A module for the idle process, a process that runs when there is nothing else left to do.

init
A module for the init process

ready_queue
A module for the ready queue

reaper
A module for the reaper process, a process which frees the resources of dead processes.

user
A module for user processes

Structs

Process
Represents a single process, its identity, resources, etc.

Enums

State
An enum representing the possible states of a process

Constants

STACK_SIZE
Size of a kernel stack (number of words)

Statics

CURRENT_PROCESS
The current running process if there is one

NEXT_ID
The next available PID

Functions

_proc_yield
The unsafe function that does the actual work of choosing the next process and switching to it.

exit
Called by the current process to exit with the given exit code

init
Initialize the process subsystem. This creates the init, idle, and reaper processes, but does not start any of them yet.

proc_yield
A safe wrapper around machine::proc_yield.

start_proc
The entry point of all processes