Struct rustcode::process::Process

pub struct Process {
    name: &'static str,
    pid: usize,
    run: fn(&Process) -> usize,
    state: State,
    stack: usize,
    kcontext: KContext,
    pub addr_space: AddressSpace,
    pub disable_cnt: usize,
    pub next_proc: *mut Process,
}

Represents a single process, its identity, resources, etc.

Fields

name

Name of the process; not necessarily unique

pid

Unique 32-bit identifier

run

The routine of the process

state

The current state of the process

stack

A pointer to the kheap-allocated stack space for this process's kernel stack. This pointer is to the bottom of the stack, not the head.

kcontext

The saved kernel context for context switching

addr_space

The virtual memory address space of the process

disable_cnt

Number of calls to interrupts::on() while this process was running Interrupts are on if disable_cnt == 0

next_proc

A link for use in process queues

Methods

impl Process

fn new(name: &'static str, run: fn(&Process) -> usize) -> *mut Process

Create a new process with the given name and routine. Because processes are a fundamental abstraction, they are too low level for me to use Rust well. For this reason, a raw pointer is returned to the process, and it is the job of the caller to arrange for the process to be reaped.

fn get_stack(&mut self)

A helper to get a kernel stack for this process

fn set_state(&mut self, s: State)

Set the state of the process to s